+++ static/js/signup.js
@@ -1,11 +1,12 @@
 $(function () {
 
     $("#signup-go").on("click", function() {
+        var base = document.getElementById("base-url").getAttribute("href").slice(0, -1);
         var email = $("#signup-email").val();
         var token = $('input[name=csrfmiddlewaretoken]').val();
 
         $.ajax({
-            url: "/accounts/signup/",
+            url: base + "/accounts/signup/",
             type: "post",
             headers: {"X-CSRFToken": token},
             data: {"identity": email},